requirements.txt
Flask==2.1.3
gunicorn==20.1.0

requests==2.28.1
structlog==22.1.0

google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-cloud-bigquery==2.34.4
